+++ lib/dart_quill_table_better.dart
@@ -6,3 +6,4 @@ export 'src/table_better/language/language.dart';
 export 'src/table_better/modules/clipboard.dart';
 export 'src/table_better/register.dart';
 export 'src/table_better/ui/cell_selection.dart';
+export 'src/table_better/ui/cell_selection_controller.dart';